FindBack_AI/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML dependencies
torch>=2.2.0
torchvision>=0.17.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
git+https://github.com/openai/CLIP.git

# Image processing
opencv-python==4.8.1.78
Pillow>=10.0.0
numpy<2.0,>=1.24.0

# Vector search
faiss-cpu>=1.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
